Title: 함수를 이용한 369 게임, Time: 7ms, Memory: 0MB, Status: Wrong Answer - CodeTree

+++ 241028/함수를 이용한 369 게임/369-games-using-functions.cpp
@@ -11,9 +11,9 @@ bool IsMagicNumber(int n) {
 
 int main() {
     int cnt = 0;
-    int num1, num2;
+    long num1, num2;
     cin >> num1 >> num2;
-    for(int i = num2; i >= num1; i--){
+    for(long i = num2; i >= num1; i--){
         if(IsMagicNumber(i)){
             cnt++;
         }
